requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
requests>=2.31.0
beautifulsoup4>=4.12.0
jinja2>=3.1.4
python-dotenv>=1.0.0
